+++ docs/changelog.md
@@ -27,6 +27,7 @@ brief notes about them may accumulate here.
 * Fixed: X10 mouse coordinate reporting encoding could produce invalid outputs for large windows. Capped coordinate values to the maximum value that is representable in UTF-8 encoding
 * Fixed: font fallback now happens asynchronously from painting [#508](https://github.com/wez/wezterm/issues/508)
 * New: added [window:get_selection_text_for_pane](config/lua/window/get_selection_text_for_pane.md) method [#575](https://github.com/wez/wezterm/issues/575)
+* Fixed: implicit hyperlink rules, word and line selection now operate on logical lines which means that they deal with wrapped lines outside of the viewport. [#408](https://github.com/wez/wezterm/issues/408)
 
 ### 20210314-114017-04b7cedd